fix: Componentização de todos os formulários de pagamento
