+++ src/app/pagamento/recarregar/mastercard/page.tsx
@@ -2,37 +2,15 @@
 import { useState } from 'react'
 import { LayoutPage } from '@/components/Shared'
 import { Button } from '@/components/Button'
-import { FormInput } from '@/components/Forms/Input'
 import { IconLeftArrow } from '@/components/Icons/iconLeftArrow'
-import usePaymentStore from '@/stores/payment.store'
 import { Title } from '@/components/Title'
 import { useRouter } from 'next/navigation'
 import CircleLoading from '@/components/CircleLoading'
+import { PaymentRechargeMastercardForm } from '@/components/Payment/Form/Recharge/index.mastercard'
 
 export default function PaymentAddMastercardPage() {
   const [isLoading, setIsLoading] = useState(false)
   const router = useRouter()
-  const [email, setEmail] = useState('')
-  const [cardNumber, setCardNumber] = useState('')
-  const [cardValidity, setCardValidity] = useState('')
-  const [cardCVC, setCardCVC] = useState('')
-  const [cardOwner, setCardOwner] = useState('')
-
-  const { paymentAdd } = usePaymentStore()
-
-  const handleOnCancel = () => {
-    router.push('/')
-  }
-
-  const validateForm = () => {
-    return (
-      email.length > 0 &&
-      cardNumber.length > 0 &&
-      cardValidity.length > 0 &&
-      cardCVC.length > 0 &&
-      cardOwner.length > 0
-    )
-  }
 
   const handleOnSubmit = (event: any) => {
     event.preventDefault()
@@ -53,7 +31,7 @@ export default function PaymentAddMastercardPage() {
             <div className="mb-8 flex w-full items-center justify-start">
               <Button
                 className="border-transparent"
-                onClick={() => handleOnCancel()}
+                onClick={() => router.push('/')}
               >
                 <IconLeftArrow />
               </Button>
@@ -69,84 +47,10 @@ export default function PaymentAddMastercardPage() {
             </div>
             <div className="flex h-full w-full flex-col items-start justify-center">
               <Title size="2xl"> Recarregar saldo com Cartão </Title>
-              <form onSubmit={() => handleOnSubmit} className="mt-4 w-full">
-                <FormInput
-                  label="Email"
-                  type="email"
-                  name="mastercard-email"
-                  placeholder="email@example.com"
-                  value={email}
-                  onChange={(event) => setEmail(event.target.value)}
+              <form className="mt-4 w-full">
+                <PaymentRechargeMastercardForm
+                  handleFormSubmit={(event) => handleOnSubmit(event)}
                 />
-
-                <FormInput
-                  label="Informações do Cartão"
-                  type="text"
-                  name="mastercard-card-number"
-                  rounded="rounded-t-md"
-                  placeholder="0000 0000 0000 0000"
-                  value={cardNumber}
-                  onChange={(event) => setCardNumber(event.target.value)}
-                />
-
-                <div className="-mt-5 grid w-full grid-cols-2 items-center">
-                  <FormInput
-                    type="text"
-                    name="mastercard-card-validity"
-                    placeholder="MM / YY"
-                    classname="border-r-0"
-                    rounded="rounded-bl-md"
-                    value={cardValidity}
-                    onChange={(event) => setCardValidity(event.target.value)}
-                  />
-                  <FormInput
-                    type="text"
-                    name="mastercard-card-cvc"
-                    placeholder="CVC"
-                    rounded="rounded-br-md"
-                    value={cardCVC}
-                    onChange={(event) => setCardCVC(event.target.value)}
-                  />
-                </div>
-
-                <FormInput
-                  label="Nome do Portador"
-                  type="text"
-                  name="mastercard-card-owner"
-                  placeholder="Nome"
-                  value={cardOwner}
-                  onChange={(event) => setCardOwner(event.target.value)}
-                />
-
-                <br />
-
-                <div className="flex justify-between text-xl font-semibold">
-                  <text>Total:</text>
-                  {/* CHANGE COLOR */}
-                  <span className="text-mesh-color-primary-800">
-                    R${paymentAdd.value?.toFixed(2).replace('.', ',')}
-                  </span>
-                </div>
-
-                {/* CHANGE COLOR */}
-                <div className="flex flex-col gap-4 text-xl font-semibold">
-                  <Button
-                    type="submit"
-                    onClick={(event) => handleOnSubmit(event)}
-                    disable={!validateForm()}
-                    className="h-12 w-full border-transparent"
-                    color="green"
-                  >
-                    Pagar
-                  </Button>
-                  <Button
-                    className="w-full border-2 py-2"
-                    onClick={() => handleOnCancel()}
-                    color="invisible"
-                  >
-                    Cancelar
-                  </Button>
-                </div>
               </form>
             </div>
           </div>
