+++ src/app/pagamento/recarregar/pix/page.tsx
@@ -2,28 +2,15 @@
 import { useState } from 'react'
 import { LayoutPage } from '@/components/Shared'
 import { Button } from '@/components/Button'
-import { FormInput } from '@/components/Forms/Input'
 import { IconLeftArrow } from '@/components/Icons/iconLeftArrow'
-import usePaymentStore from '@/stores/payment.store'
 import { Title } from '@/components/Title'
 import { useRouter } from 'next/navigation'
 import CircleLoading from '@/components/CircleLoading'
+import { PaymentRechargePixForm } from '@/components/Payment/Form/Recharge/index.pix'
 
 export default function PaymentAddPixPage() {
-  const [isLoading, setIsLoading] = useState(false)
   const router = useRouter()
-  const [name, setName] = useState('')
-  const [identification, setIdentification] = useState('')
-
-  const { paymentAdd } = usePaymentStore()
-
-  const handleOnCancel = () => {
-    router.push('/')
-  }
-
-  const validateForm = () => {
-    return name.length > 0 && identification.length > 0
-  }
+  const [isLoading, setIsLoading] = useState(false)
 
   const handleOnSubmit = (event: any) => {
     event.preventDefault()
@@ -44,7 +31,7 @@ export default function PaymentAddPixPage() {
             <div className="mb-8 flex w-full items-center justify-start">
               <Button
                 className="border-transparent"
-                onClick={() => handleOnCancel()}
+                onClick={() => router.push('/')}
               >
                 <IconLeftArrow />
               </Button>
@@ -60,54 +47,10 @@ export default function PaymentAddPixPage() {
             </div>
             <div className="flex h-full w-full flex-col items-start justify-center">
               <Title size="2xl"> Recarregar saldo com Pix </Title>
-              <form onSubmit={() => handleOnSubmit} className="mt-4 w-[546px]">
-                <FormInput
-                  label="Nome Completo"
-                  type="name"
-                  name="pix-name"
-                  placeholder="Nome e Sobrenome"
-                  value={name}
-                  onChange={(event) => setName(event.target.value)}
+              <form className="mt-4 w-[546px]">
+                <PaymentRechargePixForm
+                  handleFormSubmit={(event) => handleOnSubmit(event)}
                 />
-
-                <FormInput
-                  label="CPF/CNPJ"
-                  type="text"
-                  name="pix-cpf"
-                  placeholder="CPF ou CNPJ"
-                  value={identification}
-                  onChange={(event) => setIdentification(event.target.value)}
-                />
-
-                <br />
-
-                <div className="mt-4 flex justify-between text-xl font-semibold">
-                  <text>Total:</text>
-                  {/* CHANGE COLOR */}
-                  <span className="text-mesh-color-primary-800">
-                    R${paymentAdd.value?.toFixed(2).replace('.', ',')}
-                  </span>
-                </div>
-
-                {/* CHANGE COLOR */}
-                <div className="flex flex-col gap-4 text-xl font-semibold">
-                  <Button
-                    type="submit"
-                    onClick={(event) => handleOnSubmit(event)}
-                    disable={!validateForm()}
-                    className="h-12 w-full border-transparent"
-                    color="green"
-                  >
-                    Continuar
-                  </Button>
-                  <Button
-                    className="w-full border-2 py-2"
-                    onClick={() => handleOnCancel()}
-                    color="invisible"
-                  >
-                    Cancelar
-                  </Button>
-                </div>
               </form>
             </div>
           </div>
